# front-end/src/models/user.js
export default class User {
    constructor(name, username, email, password) {
      this.name = name
      this.username = username;
      this.email = email;
      this.password = password;
    }
  }